fix bug of switching samples cause artifact seed layers to lose event listeners
