+++ cylinter/utils.py
@@ -6,6 +6,7 @@ import pickle
 import logging
 from dataclasses import dataclass
 from typing import Dict
+from uuid import uuid4
 
 import numpy as np
 import pandas as pd
@@ -921,6 +922,45 @@ class ArtifactInfo():
                                            self.artifact_layer.data)
         self.artifact_layer.refresh()
 
+    def bind_listener_seeds(self, viewer, global_state, tolerance_spinbox):
+        seed_layer = self.seed_layer
+        if seed_layer is None:
+            return
+        def point_clicked_callback(event):
+            current_layer = viewer.layers.selection.active
+            global_state.current_layer = current_layer
+            features = seed_layer.current_properties
+            global_state.current_point = self.seeds[features['id'][0]]
+            global_state.current_tol = features['tol'][0]
+            tolerance_spinbox.value = features['tol'][0]
+        
+        def point_changed_callback(event):
+            current_layer = viewer.layers.selection.active
+            abx_channel = current_layer.metadata['abx_channel']
+            pt_id = current_layer.current_properties['id'][0]
+            artifact_info = self
+            im_transformed = artifact_info.transformed
+            global_state.current_layer = current_layer
+            df = artifact_info.seed_layer.features.copy() # preemptive...but might be wasteful if df is large
+            
+            if event.action=='add':
+                df = seed_layer.features
+                pt_id = uuid4()
+                df.loc[df.index[-1], 'id']=pt_id
+                df.loc[df.index[-1], 'tol']=0
+                seed = (current_layer.data[-1] / (2**current_layer.metadata['downscale'])).astype(int)
+                artifact_info.seeds[pt_id] = seed
+                new_fill = flood(im_transformed, seed_point=tuple(seed), 
+                                                tolerance=0)
+                artifact_info.update_mask(artifact_info.mask + new_fill)
+            elif event.action=='remove':
+                optimal_tol = global_state.current_tol
+                old_fill = flood(im_transformed, seed_point=tuple(global_state.current_point), 
+                                tolerance=optimal_tol) 
+                artifact_info.update_mask(artifact_info.mask - old_fill)
+        
+        seed_layer.events.current_properties.connect(point_clicked_callback)
+        seed_layer.events.data.connect(point_changed_callback)
     def render_seeds(self, viewer, loaded_ims, layer_name, abx_channel):
         if len(self.seeds) > 0:
             seeds = np.vstack(list(self.seeds.values()))
